requirements.txt
python-dotenv
langdetect